iot-edge/requirements.txt
opencv-python
keyboard
facenet-pytorch
torch
torchvision
pillow
numpy
scipy
qrcode
uuid
pyzbar
requests
pyttsx3
